Create README.md
Diabetes is a complex disease influenced by both genetic and environmental factors. The Pima Indians tribe presents an interesting case study as they have a high prevalence of diabetes. This report aims to explore the dataset and develop a predictive model to forecast diabetes outcomes in the tribe. Here, we are going to analyze different aspects of Diabetes in the Pima Indians tribe by doing Exploratory Data Analysis and Predictive Analysis to predict the Outcome on the basis of input features.